+++ nexus_constructor/json/load_from_json.py
@@ -1,14 +1,11 @@
 import json
-from typing import Dict, List, Optional, Tuple, Union
-
-import numpy as np
+from typing import Dict, Optional, Tuple
 
 from nexus_constructor.common_attrs import (
     PIXEL_SHAPE_GROUP_NAME,
     SHAPE_GROUP_NAME,
     CommonAttrs,
     CommonKeys,
-    NodeType,
 )
 from nexus_constructor.component_type import COMPONENT_TYPES
 from nexus_constructor.json.json_warnings import (
@@ -20,7 +17,11 @@ from nexus_constructor.json.json_warnings import (
 )
 from nexus_constructor.json.load_from_json_utils import (
     DEPENDS_ON_IGNORE,
+    _add_field_to_group,
+    _find_depends_on_path,
     _find_nx_class,
+    _find_shape_information,
+    _retrieve_children_list,
 )
 from nexus_constructor.json.shape_reader import ShapeReader
 from nexus_constructor.json.transform_id import TransformId
@@ -29,40 +30,10 @@ from nexus_constructor.json.transformation_reader import (
     get_component_and_transform_name,
 )
 from nexus_constructor.model.component import Component
-from nexus_constructor.model.dataset import Dataset
 from nexus_constructor.model.entry import Entry
-from nexus_constructor.model.group import TRANSFORMS_GROUP_NAME, Group
+from nexus_constructor.model.group import TRANSFORMS_GROUP_NAME
 from nexus_constructor.model.instrument import Instrument
-from nexus_constructor.model.stream import (
-    ADC_PULSE_DEBUG,
-    ARRAY_SIZE,
-    CHUNK_CHUNK_KB,
-    CHUNK_CHUNK_MB,
-    DATA_TYPE,
-    DATASET,
-    EDGE_TYPE,
-    ERROR_TYPE,
-    INDEX_EVERY_KB,
-    INDEX_EVERY_MB,
-    LINK,
-    SHAPE,
-    SOURCE,
-    STORE_LATEST_INTO,
-    TOPIC,
-    VALUE_UNITS,
-    EV42Stream,
-    F142Stream,
-    HS00Stream,
-    Link,
-    NS10Stream,
-    SENVStream,
-    Stream,
-    StreamGroup,
-    TDCTStream,
-    WriterModules,
-)
 from nexus_constructor.model.transformation import Transformation
-from nexus_constructor.model.value_type import VALUE_TYPE_TO_NP
 
 """
 The current implementation makes a couple of assumptions that may not hold true for all valid JSON descriptions of
@@ -78,182 +49,6 @@ CHILD_EXCLUDELIST = [
     TRANSFORMS_GROUP_NAME,
     CommonAttrs.DEPENDS_ON,
 ]
-
-
-def _retrieve_children_list(json_dict: Dict) -> List:
-    """
-    Attempts to retrieve the children from the JSON dictionary.
-    :param json_dict: The JSON dictionary loaded by the user.
-    :return: The children value is returned if it was found, otherwise an empty list is returned.
-    """
-    value = []
-    try:
-        entry = json_dict[CommonKeys.CHILDREN][0]
-        value = entry[CommonKeys.CHILDREN]
-    except (KeyError, IndexError, TypeError):
-        pass
-    return value
-
-
-def _find_shape_information(children: List[Dict]) -> Union[Dict, None]:
-    """
-    Tries to get the shape information from a component.
-    :param children: The list of dictionaries.
-    :return: The shape attribute if it could be found, otherwise None.
-    """
-    value = None
-    try:
-        for item in children:
-            if item[CommonKeys.NAME] in [SHAPE_GROUP_NAME, PIXEL_SHAPE_GROUP_NAME]:
-                value = item
-    except KeyError:
-        pass
-    return value
-
-
-def _add_field_to_group(item: Dict, group: Group):
-    child: Group
-    if CommonKeys.TYPE in item:
-        if item[CommonKeys.NAME] == TRANSFORMS_GROUP_NAME:
-            return
-        field_type = item[CommonKeys.TYPE]
-        child_name = item[CommonKeys.NAME]
-        if field_type == NodeType.GROUP:
-            child = _create_group(item, group)
-        else:
-            raise Exception(
-                f'Found unknown field type ("{field_type}") when loading JSON - {child_name}'
-            )
-        group[child_name] = child
-    elif CommonKeys.MODULE in item:
-        stream: Union[
-            Dataset,
-            Link,
-            NS10Stream,
-            SENVStream,
-            TDCTStream,
-            EV42Stream,
-            F142Stream,
-            HS00Stream,
-        ]
-        writer_module = item[CommonKeys.MODULE]
-        if writer_module == LINK:
-            stream = _create_link(item)
-        elif writer_module == DATASET:
-            if item[NodeType.CONFIG][CommonKeys.NAME] == CommonAttrs.DEPENDS_ON:
-                return
-            stream = _create_dataset(item, group)
-        else:
-            stream = _create_stream(item)
-        group.children.append(
-            stream
-        )  # Can't use the `[]` operator because streams do not have a name to use as a key
-    else:
-        raise Exception(
-            "Unable to add field as neither writer module type nor child type was found in the current node."
-        )
-
-
-def _find_depends_on_path(items: List[Dict], name: str) -> Optional[str]:
-    if not isinstance(items, list):
-        raise RuntimeError(
-            f'List of children in node with the name "{name}" is not a list.'
-        )
-    for item in items:
-        try:
-            config = item[NodeType.CONFIG]
-            if config[CommonKeys.NAME] != CommonAttrs.DEPENDS_ON:
-                continue
-            return config[CommonKeys.VALUES]
-        except KeyError:
-            pass  # Not all items has a config node, ignore those that do not.
-    return None
-
-
-def _create_stream(json_object: Dict) -> Stream:
-    """
-    Given a dictionary containing a stream, create a corresponding stream object to be used in the model.
-    :param json_object: JSON dictionary containing a stream.
-    :return: A stream object containing relevant data from the JSON.
-    """
-    stream_object = json_object[NodeType.CONFIG]
-    writer_module = json_object[CommonKeys.MODULE]
-    source = stream_object[SOURCE]
-    topic = stream_object[TOPIC]
-    # Common to ev42 and f142 stream objects
-    index_mb = (
-        stream_object[INDEX_EVERY_MB] if INDEX_EVERY_MB in stream_object else None
-    )
-    index_kb = (
-        stream_object[INDEX_EVERY_KB] if INDEX_EVERY_KB in stream_object else None
-    )
-    if writer_module == WriterModules.F142.value:
-        return __create_f142_stream(index_kb, index_mb, source, stream_object, topic)
-    if writer_module == WriterModules.EV42.value:
-        return __create_ev42_stream(index_kb, index_mb, source, stream_object, topic)
-    if writer_module == WriterModules.HS00.value:
-        data_type = stream_object[DATA_TYPE]
-        error_type = stream_object[ERROR_TYPE]
-        edge_type = stream_object[EDGE_TYPE]
-        shape = stream_object[SHAPE]
-        return HS00Stream(
-            source=source,
-            topic=topic,
-            data_type=data_type,
-            error_type=error_type,
-            edge_type=edge_type,
-            shape=shape,
-        )
-    if writer_module == WriterModules.NS10.value:
-        return NS10Stream(source=source, topic=topic)
-    if writer_module == WriterModules.SENV.value:
-        return SENVStream(source=source, topic=topic)
-    if writer_module == WriterModules.TDCTIME.value:
-        return TDCTStream(source=source, topic=topic)
-
-    return None
-
-
-def __create_ev42_stream(
-    index_kb: str, index_mb: str, source: str, stream_object: Dict, topic: str
-):
-    adc = stream_object[ADC_PULSE_DEBUG] if ADC_PULSE_DEBUG in stream_object else None
-    chunk_mb = (
-        stream_object[CHUNK_CHUNK_MB] if CHUNK_CHUNK_MB in stream_object else None
-    )
-    chunk_kb = (
-        stream_object[CHUNK_CHUNK_KB] if CHUNK_CHUNK_KB in stream_object else None
-    )
-    return EV42Stream(
-        source=source,
-        topic=topic,
-        adc_pulse_debug=adc,
-        nexus_indices_index_every_kb=index_kb,
-        nexus_indices_index_every_mb=index_mb,
-        nexus_chunk_chunk_mb=chunk_mb,
-        nexus_chunk_chunk_kb=chunk_kb,
-    )
-
-
-def __create_f142_stream(
-    index_kb: str, index_mb: str, source: str, stream_object: Dict, topic: str
-):
-    value_type = stream_object[CommonKeys.DATA_TYPE]
-    value_units = stream_object[VALUE_UNITS] if VALUE_UNITS in stream_object else None
-    array_size = stream_object[ARRAY_SIZE] if ARRAY_SIZE in stream_object else None
-    store_latest_into = (
-        stream_object[STORE_LATEST_INTO] if STORE_LATEST_INTO in stream_object else None
-    )
-    return F142Stream(
-        source=source,
-        topic=topic,
-        type=value_type,
-        value_units=value_units,
-        array_size=array_size,
-        nexus_indices_index_every_mb=index_mb,
-        nexus_indices_index_every_kb=index_kb,
-        store_latest_into=store_latest_into,
-    )
 
 
 class JSONReader:
@@ -455,58 +250,3 @@ class JSONReader:
             return False
 
         return True
-
-
-def _create_group(json_object: Dict, parent: Group) -> Group:
-    children = json_object[CommonKeys.CHILDREN]
-    name = json_object[CommonKeys.NAME]
-    group = Group(name=name, parent_node=parent)
-    for item in children:
-        if CommonKeys.MODULE in item:
-            group = StreamGroup(name=name, parent_node=parent)
-            break
-
-    for item in children:
-        _add_field_to_group(item, group)
-
-    _add_attributes(json_object, group)
-    return group
-
-
-def _get_data_type(json_object: Dict):
-    if CommonKeys.DATA_TYPE in json_object:
-        return json_object[CommonKeys.DATA_TYPE]
-    elif CommonKeys.TYPE in json_object:
-        return json_object[CommonKeys.TYPE]
-    raise KeyError
-
-
-def _create_dataset(json_object: Dict, parent: Group) -> Dataset:
-    value_type = _get_data_type(json_object[NodeType.CONFIG])
-    name = json_object[NodeType.CONFIG][CommonKeys.NAME]
-    values = json_object[NodeType.CONFIG][CommonKeys.VALUES]
-    if isinstance(values, list):
-        # convert to a numpy array using specified type
-        values = np.array(values, dtype=VALUE_TYPE_TO_NP[value_type])
-    ds = Dataset(name=name, values=values, type=value_type, parent_node=parent)
-    _add_attributes(json_object, ds)
-    return ds
-
-
-def _create_link(json_object: Dict) -> Link:
-    name = json_object[NodeType.CONFIG][CommonKeys.NAME]
-    target = json_object[NodeType.CONFIG][SOURCE]
-    return Link(name=name, target=target)
-
-
-def _add_attributes(json_object: Dict, model_object: Union[Group, Dataset]):
-    try:
-        attrs_list = json_object[CommonKeys.ATTRIBUTES]
-        for attribute in attrs_list:
-            attr_name = attribute[CommonKeys.NAME]
-            attr_values = attribute[CommonKeys.VALUES]
-            model_object.attributes.set_attribute_value(
-                attribute_name=attr_name, attribute_value=attr_values
-            )
-    except (KeyError, AttributeError):
-        pass
