+++ tests/json/test_shape_reader.py
@@ -15,9 +15,9 @@ from nexus_constructor.common_attrs import (
     NodeType,
 )
 from nexus_constructor.json.json_warnings import JsonWarningsContainer
-from nexus_constructor.json.load_from_json import _get_data_type
 from nexus_constructor.json.load_from_json_utils import (
     _find_attribute_from_list_or_dict,
+    _get_data_type,
 )
 from nexus_constructor.json.shape_reader import (
     DETECTOR_NUMBER,
